+++ packages/css/src/types.ts
@@ -605,20 +605,6 @@ export interface ThemeStyles {
   [key: string]: ThemeUIStyleObject | undefined
 }
 
-/**
- * @internal
- * We fall back to `theme.useBodyStyles` when `theme.useRootStyles` is not set.
- */
-export function __internalGetUseRootStyles(theme: Theme = {}) {
-  const { useRootStyles = {}, useBodyStyles = {} } = theme
-  const root = 'useRootStyles' in theme && useRootStyles != null
-
-  return {
-    scope: root ? 'html' : 'body',
-    rootStyles: root ? useRootStyles : useBodyStyles,
-  }
-}
-
 export interface Theme {
   breakpoints?: Array<string>
   mediaQueries?: { [size: string]: string }