+++ packages/core/src/index.ts
@@ -32,6 +32,9 @@ export type {
   ThemeUIStyleObject,
   VariantProperty,
 } from '@theme-ui/css'
+
+
+export { __internalGetUseRootStyles } from '@theme-ui/css'
 export * from './types'
 
 const __EMOTION_VERSION__ = packageInfo.version